+++ src/components/Content.tsx
@@ -1,4 +1,4 @@
-import { NavTopics } from "./NavTopics";
+import { Sidebar } from "./Sidebar";
 import { CreateNote } from "./CreateNote";
 import { FaInfoCircle } from "react-icons/fa";
 import { useSelectedTopic } from "~/contexts/SelectedTopicContext";
@@ -15,7 +15,7 @@ export const Content: React.FC = () => {
   };
   return (
     <div className="mx-5 mt-5 flex min-h-[75vh] flex-col md:flex-row">
-      <NavTopics />
+        <Sidebar />
       <div className="hide mt-8 inline-block min-h-[1em] w-[1px] self-stretch bg-gray-300 opacity-100 dark:opacity-25"></div>
       {selectedTopicId !== null ? <CreateNote /> : <NotSelectedTopic />}
     </div>
