+++ src/styles/globals.css
@@ -5,7 +5,7 @@
 @media only screen and (max-width: 640px) and (min-width: 280px) {
     .order {
         display: flex;
-        flex-direction: column-reverse;
+        flex-direction: column-reverse!important;
     }
 
     .hide {
